refactor(components): rehome the misnamed aos4 component folder
`src/components/aos4/` was named for the game edition rather than for what it
held, and it held two unrelated things: a shared card header and a hook.

The naming had a real consequence. `legacyIsolation.test.ts` checks every
presentation file against an explicit allowlist, but it only walks a fixed list
of `presentationRoots` -- `components/helpers`, `components/info`,
`components/input`, and so on. `components/aos4` was not among them, so both
files sat outside the boundary check entirely.

Each file moves to the directory that matches what it actually is:

- `collapsibleCardHeader.tsx` -> `components/helpers/`, alongside the other
  shared UI helpers. It is used by both `info/reminders.tsx` and
  `input/army_builder.tsx`.
- `useIsMobile.ts` -> `utils/hooks/`, joining `useInterval`, `useLogin`, and
  `useWindowSize`. It is a hook, not a component.

The card header is added to the isolation allowlist, so it is now genuinely
covered by that check rather than invisible to it. No behavior changes; the
edits are import paths only.
