+++ src/components/input/army_builder.tsx
@@ -1,7 +1,7 @@
 import type { CanonicalId } from '../../aos4/domain'
 import type { createAos4BuilderViewModel } from '../../aos4/view'
-import { CollapsibleCardHeader } from 'components/aos4/collapsibleCardHeader'
-import { useIsMobile } from 'components/aos4/useIsMobile'
+import { CollapsibleCardHeader } from 'components/helpers/collapsibleCardHeader'
+import { useIsMobile } from 'utils/hooks/useIsMobile'
 import { useTheme } from 'context/useTheme'
 import { useMemo, useState } from 'react'
 import Select, { type MultiValue } from 'react-select'
